+++ src/pathanimations/astar/heuristics/Manhattan.java
@@ -5,7 +5,7 @@ import graph.Point;
 public class Manhattan implements Heuristic {
 
     @Override
-    public float getFor(Point p1, Point p2) {
+    public double getFor(Point p1, Point p2) {
         return Math.abs(p1.x() - p2.x()) + Math.abs(p1.y() - p2.y());
     }
 }
